Refactor : searchResult 컴포넌트 분리
